Attach character creation attribute inputs to redux [Re: nmkataoka/adv-life#79]
